Refactor: Modularize `src/routes/index.tsx` into smaller components
