+++ src/components/sections/philosophy.tsx
@@ -1,0 +1,14 @@
+export function Philosophy() {
+  return (
+    <section className="hairline-t">
+      <div className="container-page py-24 md:py-36">
+        <div className="eyebrow">Leadership Philosophy</div>
+        <blockquote className="mt-6 display-2 max-w-4xl">
+          “Technology succeeds when strategy, engineering, and execution operate as a
+          <em className="italic text-muted-foreground"> single system.</em> My role is to create
+          clarity, alignment, and predictable outcomes — while enabling highly autonomous teams.”
+        </blockquote>
+      </div>
+    </section>
+  );
+}
