+++ src/components/sections/vision.tsx
@@ -1,0 +1,18 @@
+export function Vision() {
+  return (
+    <section className="hairline-t bg-surface">
+      <div className="container-page py-24 md:py-32 grid gap-10 md:grid-cols-[1fr_2fr] md:gap-20">
+        <div>
+          <div className="eyebrow">Vision</div>
+          <h2 className="display-2 mt-4">The next five years.</h2>
+        </div>
+        <p className="text-lg md:text-xl leading-relaxed text-muted-foreground max-w-2xl">
+          Evolving from delivery leadership into large-scale{" "}
+          <span className="text-foreground">Technical Program Management</span> — leading strategic
+          programs that combine software, AI, cloud infrastructure, and emerging technologies at
+          global scale.
+        </p>
+      </div>
+    </section>
+  );
+}
